Rosewood- Updates the Town Hotspring (Again) (#1539)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

This PR makes a dedicated walkable path through the hotspring to
potentially help with traffic and navigation


![PR](https://github.com/user-attachments/assets/408872cb-4f12-48f1-b23b-095132834cbe)

This also fixes a few misplaced missing areas, fills the rivers with
plants, and makes beggar hell not lethal to someone who is just floating
down the river


## Why It's Good For The Game

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->
Map QOL good

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->

---------

Co-authored-by: tonty <39193182+tontyGH@users.noreply.github.com>