Youngling werewolf sprite (#1487)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

* Adds sprites for youngling werewolvs + wound overlays

## Why It's Good For The Game

Its neat to have uniqe sprites for young werewolvs.
Its not too diffrent that you can immediatly tell its a youngling, but
distinct enough that there is a diffrence.
Similar to how there is a diffrence between male and female werewolfs.

Sprites:
https://imgur.com/a/UQhX7Xc

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->