Decorated Wall Smoothening, + Erebus + Floor types. (#1579)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request
-Adds a new type of decorated wall that has smoothening.
-Adds new subtypes of said decorated wall that has smoothening.
Note: This does not adjust existing decorated walls, as I do not want to
map conflict to shit right now. This will be phased out on a later date.
-Changes erebus walls to utilize smoothening.
-Adds new purple variants of church tiles (At Drag's request.)

![SmoothWalls!!](https://github.com/user-attachments/assets/7dbd6ce9-f08f-40ac-950e-cff5e734ec6f)

![SmoothWalls!!_Underworld](https://github.com/user-attachments/assets/d454fd2e-b75d-4285-a06c-c28a900b4bb8)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

## Why It's Good For The Game
I've been mapping Daftmarsh for almost a month now, and the amount of
hair I have wanted to pull from my scalp due to the manual smoothening
you need to employ for the church/keep/anything that uses that type of
wall is insane.
Also, more tile variants for mapper creativity.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->

---------

Co-authored-by: AberrantQuerist <crystalseeker04@gmail.com>