Boat QOL and Refactor, Merchant faction buffs (#1483)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

- Refactors boat to act more like SS13 cargo, overall QOL in the cargo
ordering and selling process
- Adds shophand as a merchant assistant. Gives them dock access and
removes gaffer access.
- Stevedores buffed in gear and skills, CRATEMOVER trait now gives the
strong pull component
- Department flags changed to move Pilgrim, Adventurer, and Mercenary
into an "Outsiders" category for the latejoin menu. Merchant, Shophand,
and Stevedore moved into "Company" category and `GLOB.company_positions`

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

## Why It's Good For The Game
Long live the Company.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->