Ports ratwood infernal feather behavior, makes it actually work. (#1568)
## About The Pull Request
Ports paper & infernal feather behavior from
https://github.com/Rotwood-Vale/Ratwood-Keep/pull/2213
Currently infernal feather was ported but had no paper code ported with
it, so it was not doing anything at all.

I also made it so you don't need to just use a use intent for it to
activate, but you can also click read in the chat bar for it to
activate.
## Why It's Good For The Game

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->
Makes the infernal feather actually do something at all.
## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->